Move most of ScheduledMail back to internal
JIRA: RAIL-2956
