+++ libs/sdk-ui-ext/src/dashboardView/ScheduledMailDialog.tsx
@@ -8,20 +8,19 @@ import {
     IFilterContextDefinition,
     FilterContextItem,
 } from "@gooddata/sdk-backend-spi";
-import { ObjRef } from "@gooddata/sdk-model";
+import { ObjRef, uriRef } from "@gooddata/sdk-model";
 import { GoodDataSdkError, LoadingComponent, ErrorComponent } from "@gooddata/sdk-ui";
 import { Overlay } from "@gooddata/sdk-ui-kit";
 
-import { useDashboard } from "../../hooks";
+import { useDashboard } from "./hooks";
 import {
     useCurrentUser,
     useSaveScheduledMail,
     useUserWorkspacePermissions,
     useUserWorkspaceSettings,
-} from "../../hooks/internal";
+} from "./hooks/internal";
 
-import { ScheduledMailDialogRenderer } from "./ScheduledMailDialogRenderer";
-import { uriRef } from "@gooddata/sdk-model";
+import { ScheduledMailDialogRenderer } from "../internal";
 
 export type ScheduledMailDialogProps = {
     /**
