+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/ScheduledMailDialog/RecipientsSelect/RecipientsSelectRenderer.tsx
