+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/ScheduledMailDialog/RepeatSelect/RepeatExecuteOnSelect.tsx
