+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/ScheduledMailDialog/RepeatSelect/RepeatFrequencySelect.tsx
