+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/ScheduledMailDialog/RepeatSelect/test/RepeatExecuteOnSelect.test.tsx
@@ -6,7 +6,7 @@ import { mount, ReactWrapper } from "enzyme";
 import { REPEAT_EXECUTE_ON } from "../../../constants";
 import { getDate, getDayName, getWeek } from "../../../utils/datetime";
 import { RepeatExecuteOnSelect, IRepeatExecuteOnSelectProps } from "../RepeatExecuteOnSelect";
-import { InternalIntlWrapper } from "../../../../../internal";
+import { InternalIntlWrapper } from "../../../../../utils/internalIntlProvider";
 
 import {
     getDropdownTitle,
