+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/ScheduledMailDialog/RepeatSelect/test/RepeatTypeSelect.test.tsx
@@ -6,7 +6,7 @@ import noop from "lodash/noop";
 import { RepeatTypeSelect, IRepeatTypeSelectProps } from "../RepeatTypeSelect";
 import { getDayName, getWeek } from "../../../utils/datetime";
 import { REPEAT_TYPES } from "../../../constants";
-import { InternalIntlWrapper } from "../../../../../internal";
+import { InternalIntlWrapper } from "../../../../../utils/internalIntlProvider";
 
 import {
     getDropdownTitle,
