+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/ScheduledMailDialog/ScheduledMailDialogRenderer.tsx
@@ -40,7 +40,7 @@ import { Input } from "./Input";
 import { DateTime } from "./DateTime";
 import { Attachment } from "./Attachment";
 import { RecipientsSelect } from "./RecipientsSelect/RecipientsSelect";
-import { InternalIntlWrapper } from "../../../internal";
+import { InternalIntlWrapper } from "../../../utils/internalIntlProvider";
 
 const MAX_MESSAGE_LENGTH = 200;
 const MAX_SUBJECT_LENGTH = 200;
