+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/ScheduledMailDialog/test/Textarea.test.tsx
@@ -3,22 +3,25 @@ import React from "react";
 import { mount, ReactWrapper } from "enzyme";
 import noop from "lodash/noop";
 
-import { Input, IInputProps } from "../Input";
-import { InternalIntlWrapper } from "../../../../internal";
+import { Textarea, ITextareaProps } from "../Textarea";
+import { InternalIntlWrapper } from "../../../../utils/internalIntlProvider";
 
-describe("Input", () => {
-    function renderComponent(customProps: Partial<IInputProps> = {}): ReactWrapper {
+describe("Textarea", () => {
+    function renderComponent(customProps: Partial<ITextareaProps> = {}): ReactWrapper {
         const defaultProps = {
+            hasError: false,
+            hasWarning: false,
             label: "",
             maxlength: 100,
             placeholder: "",
+            rows: 4,
             onChange: noop,
             ...customProps,
         };
 
         return mount(
             <InternalIntlWrapper>
-                <Input {...defaultProps} />
+                <Textarea {...defaultProps} />
             </InternalIntlWrapper>,
         );
     }
@@ -32,5 +35,15 @@ describe("Input", () => {
         const label = "subject";
         const component = renderComponent({ label });
         expect(component.find("label.gd-label").text()).toBe(label);
+    });
+
+    it("should trigger onChange event", () => {
+        const value = "new value";
+        const onChange = jest.fn();
+        const component = renderComponent({ onChange });
+
+        component.find("textarea").simulate("change", { target: { value } });
+
+        expect(onChange).toBeCalledWith(value);
     });
 });
