+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/constants.ts
