+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/index.ts
@@ -7,5 +7,3 @@ export {
     ScheduledMailDialogRenderer,
     IScheduledMailDialogRendererOwnProps,
 } from "./ScheduledMailDialog/ScheduledMailDialogRenderer";
-
-export { ScheduledMailDialog } from "./ScheduledMailDialog/ScheduledMailDialog";
