+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/interfaces.ts
