+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/utils/scheduledMailSummary.ts
