+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/ScheduledMail/utils/test/validate.test.ts
